requirements_clearlinux.txt
google-cloud-storage==1.20.0
falcon==2.0.0
cheroot==8.2.1
tensorflow-serving-api==2.0.0